requirements.txt
feedparser==6.0.10
sodapy==2.2.0
pandas==2.2.0
atproto==0.0.63
requests==2.32.5
python-dotenv==1.2.1
kafka-python==2.3.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Spark Streaming & Deduplication (Ticket 3.2)
pyspark==3.5.3
sentence-transformers==5.1.2
scikit-learn==1.7.2

# Ticket 4.1 (S3 Upload)
boto3
botocore
praw
# Location Extraction & Geospatial (Ticket 3.3)
spacy==3.8.3
geopy==2.4.1
shapely==2.0.6

numpy>=1.24.0,<2.0.0
